agora/cerebral_rag/src/requirements.txt
chromadb==0.4.22
langchain==0.1.9
langchain-community==0.0.24
python-magic==0.4.27
pypdf==4.0.1
tensorflow-text==2.14.0
tensorflow-hub==0.15.0
numpy==1.24.3
pandas==2.1.4
tqdm==4.66.1
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
onnxruntime-genai
huggingface-hub[cli]
flask 
flask-socketio 
eventlet